requirements.txt
### Framework requirements
############################

python-magic
PrettyTable
requests
SQLAlchemy
psycopg2

### Module requirements
#########################

pycrypto
#pysqlcipher
olefile
oletools
BeautifulSoup4
bottle
pylzma
pyelftools
bitstring
dnspython
git+https://github.com/smarnach/pyexiftool.git#egg=pyexiftool
pyasn1
M2Crypto
pymisp
django
pypssl
pypdns
scandir
pype32
pbkdf2
pydeep
http://pefile.googlecode.com/files/pefile-1.2.10-139.tar.gz#egg=pefile

### Integration requirements
##############################

# Carbonblack
simplejson
pyaml

### Script requirements
#########################